services/product-segmentor/requirements.txt
# Core dependencies
asyncio-mqtt==0.16.1
pydantic==2.5.0
asyncpg==0.29.0
aio-pika==9.3.1
python-dotenv>=1.0.0

# Computer vision and ML
torchvision>=0.15.0
transformers>=4.35.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0